+++ client/lib/ffi.rs
@@ -53,6 +53,7 @@ pub enum casper_error_t {
     CASPER_FFI_PTR_NULL_BUT_REQUIRED = -22,
     CASPER_CONFLICTING_ARGUMENTS = -23,
     CASPER_DEPLOY_SIZE_TOO_LARGE = -24,
+    CASPER_FAILED_TO_CREATE_DICTIONARY_IDENTIFIER = -25,
 }
 
 trait AsFFIError {
@@ -86,6 +87,9 @@ impl AsFFIError for Error {
             Error::FFIPtrNullButRequired(_) => casper_error_t::CASPER_FFI_PTR_NULL_BUT_REQUIRED,
             Error::ConflictingArguments { .. } => casper_error_t::CASPER_CONFLICTING_ARGUMENTS,
             Error::DeploySizeTooLarge(_) => casper_error_t::CASPER_DEPLOY_SIZE_TOO_LARGE,
+            Error::FailedToParseDictionaryIdentifier => {
+                casper_error_t::CASPER_FAILED_TO_CREATE_DICTIONARY_IDENTIFIER
+            }
         }
     }
 }
@@ -279,7 +283,8 @@ pub extern "C" fn casper_put_deploy(
             deploy_params,
             session_params,
             payment_params,
-        );
+        )
+        .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -350,7 +355,7 @@ pub extern "C" fn casper_send_deploy_file(
     let input_path = try_unsafe_arg!(input_path);
     runtime.block_on(async move {
         let result =
-            super::send_deploy_file(maybe_rpc_id, node_address, verbosity_level, input_path);
+            super::send_deploy_file(maybe_rpc_id, node_address, verbosity_level, input_path).await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -392,7 +397,8 @@ pub extern "C" fn casper_transfer(
             transfer_id,
             deploy_params,
             payment_params,
-        );
+        )
+        .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -449,7 +455,8 @@ pub extern "C" fn casper_get_deploy(
     let node_address = try_unsafe_arg!(node_address);
     let deploy_hash = try_unsafe_arg!(deploy_hash);
     runtime.block_on(async move {
-        let result = super::get_deploy(maybe_rpc_id, node_address, verbosity_level, deploy_hash);
+        let result =
+            super::get_deploy(maybe_rpc_id, node_address, verbosity_level, deploy_hash).await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -474,7 +481,8 @@ pub extern "C" fn casper_get_block(
     let node_address = try_unsafe_arg!(node_address);
     let maybe_block_id = try_unsafe_arg!(maybe_block_id);
     runtime.block_on(async move {
-        let result = super::get_block(maybe_rpc_id, node_address, verbosity_level, maybe_block_id);
+        let result =
+            super::get_block(maybe_rpc_id, node_address, verbosity_level, maybe_block_id).await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -483,7 +491,7 @@ pub extern "C" fn casper_get_block(
 
 /// Retrieves all `Transfer` items for a `Block` from the network.
 ///
-/// See [super::casper_get_block_transfers](super::casper_get_block_transfers) for more details.
+/// See [super::get_block_transfers](super::get_block_transfers) for more details.
 #[no_mangle]
 pub extern "C" fn casper_get_block_transfers(
     maybe_rpc_id: *const c_char,
@@ -500,7 +508,8 @@ pub extern "C" fn casper_get_block_transfers(
     let maybe_block_id = try_unsafe_arg!(maybe_block_id);
     runtime.block_on(async move {
         let result =
-            super::get_block_transfers(maybe_rpc_id, node_address, verbosity_level, maybe_block_id);
+            super::get_block_transfers(maybe_rpc_id, node_address, verbosity_level, maybe_block_id)
+                .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -526,7 +535,8 @@ pub extern "C" fn casper_get_state_root_hash(
     let maybe_block_id = try_unsafe_arg!(maybe_block_id);
     runtime.block_on(async move {
         let result =
-            super::get_state_root_hash(maybe_rpc_id, node_address, verbosity_level, maybe_block_id);
+            super::get_state_root_hash(maybe_rpc_id, node_address, verbosity_level, maybe_block_id)
+                .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -562,7 +572,8 @@ pub extern "C" fn casper_get_item(
             state_root_hash,
             key,
             path,
-        );
+        )
+        .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -595,7 +606,8 @@ pub extern "C" fn casper_get_balance(
             verbosity_level,
             state_root_hash,
             purse,
-        );
+        )
+        .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -625,7 +637,8 @@ pub extern "C" fn casper_get_era_info_by_switch_block(
             node_address,
             verbosity_level,
             maybe_block_id,
-        );
+        )
+        .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
@@ -651,11 +664,28 @@ pub extern "C" fn casper_get_auction_info(
     let maybe_block_id = try_unsafe_arg!(maybe_block_id);
     runtime.block_on(async move {
         let result =
-            super::get_auction_info(maybe_rpc_id, node_address, verbosity_level, maybe_block_id);
+            super::get_auction_info(maybe_rpc_id, node_address, verbosity_level, maybe_block_id)
+                .await;
         let response = try_unwrap_rpc!(result);
         copy_str_to_buf(&response, response_buf, response_buf_len);
         casper_error_t::CASPER_SUCCESS
     })
+}
+
+/// Generates key files.
+///
+/// See [super::keygen::generate_files](super::keygen::generate_files) for more details.
+#[no_mangle]
+pub extern "C" fn casper_keygen(
+    output_dir: *const c_char,
+    algorithm: *const c_char,
+    force: bool,
+) -> casper_error_t {
+    let output_dir = try_unsafe_arg!(output_dir);
+    let algorithm = try_unsafe_arg!(algorithm);
+    let result = super::keygen::generate_files(output_dir, algorithm, force);
+    try_unwrap_result!(result);
+    casper_error_t::CASPER_SUCCESS
 }
 
 /// Container for `Deploy` construction options.
@@ -693,8 +723,8 @@ impl TryInto<super::DeployStrParams<'_>> for casper_deploy_params_t {
             timestamp,
             ttl,
             gas_price,
-            chain_name,
             dependencies,
+            chain_name,
         })
     }
 }