+++ client/lib/parsing.rs
@@ -348,12 +348,12 @@ pub(super) fn parse_deploy_params(
     let chain_name = chain_name.to_string();
 
     Ok(DeployParams {
+        secret_key,
         timestamp,
         ttl,
         gas_price,
         dependencies,
         chain_name,
-        secret_key,
     })
 }
 
@@ -595,31 +595,16 @@ pub(crate) fn transfer_id(value: &str) -> Result<u64> {
 
 #[cfg(test)]
 mod tests {
-    use std::{convert::TryFrom, result::Result as StdResult};
-
     use casper_types::{
         account::AccountHash, bytesrepr::ToBytes, AccessRights, CLTyped, CLValue, NamedArg,
         PublicKey, RuntimeArgs, URef, U128, U256, U512,
     };
+    use std::{convert::TryFrom, io::Write, result::Result as StdResult};
+    use tempfile::tempdir;
 
     use crate::{PaymentStrParams, SessionStrParams};
 
     use super::*;
-
-    #[derive(Debug)]
-    struct ErrWrapper(pub Error);
-
-    impl PartialEq for ErrWrapper {
-        fn eq(&self, other: &ErrWrapper) -> bool {
-            format!("{:?}", self.0) == format!("{:?}", other.0)
-        }
-    }
-
-    impl From<Error> for ErrWrapper {
-        fn from(error: Error) -> Self {
-            ErrWrapper(error)
-        }
-    }
 
     mod bad {
         pub const EMPTY: &str = "";
@@ -664,20 +649,17 @@ mod tests {
         pub const ENTRY_POINT: &str = "entrypoint";
         pub const VERSION: &str = "1.0.0";
         pub const TRANSFER: bool = true;
+        pub const PRIVATE_KEY: &str = "-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEIFAr+JLnFaRwpqsAEbcYLfaDixKHGdBfFsPrLKS9VTMH\n-----END PRIVATE KEY-----";
     }
 
     fn invalid_simple_args_test(cli_string: &str) {
         assert!(
-            arg_simple::payment::parse(&[cli_string])
-                .map_err(ErrWrapper)
-                .is_err(),
+            arg_simple::payment::parse(&[cli_string]).is_err(),
             "{} should be an error",
             cli_string
         );
         assert!(
-            arg_simple::session::parse(&[cli_string])
-                .map_err(ErrWrapper)
-                .is_err(),
+            arg_simple::session::parse(&[cli_string]).is_err(),
             "{} should be an error",
             cli_string
         );
@@ -869,7 +851,7 @@ mod tests {
 
     #[test]
     fn should_fail_to_parse_conflicting_arg_types() {
-        assert_eq!(
+        assert!(matches!(
             parse_session_info(
                 "",
                 "name",
@@ -881,16 +863,14 @@ mod tests {
                 "",
                 "entrypoint",
                 false
-            )
-            .map(|_| ())
-            .map_err(ErrWrapper),
+            ),
             Err(Error::ConflictingArguments {
                 context: "parse_session_info",
-                args: vec!["session_args".to_owned(), "session_args_complex".to_owned()]
-            }
-            .into())
-        );
-        assert_eq!(
+                ..
+            })
+        ));
+
+        assert!(matches!(
             parse_payment_info(
                 "",
                 "name",
@@ -902,20 +882,17 @@ mod tests {
                 "path_to/file",
                 "",
                 "entrypoint",
-            )
-            .map(|_| ())
-            .map_err(ErrWrapper),
+            ),
             Err(Error::ConflictingArguments {
                 context: "parse_payment_info",
-                args: vec!["payment_args".to_owned(), "payment_args_complex".to_owned()]
-            }
-            .into())
-        );
+                ..
+            })
+        ));
     }
 
     #[test]
     fn should_fail_to_parse_conflicting_session_parameters() {
-        assert_eq!(
+        assert!(matches!(
             parse_session_info(
                 happy::HASH,
                 happy::NAME,
@@ -927,42 +904,145 @@ mod tests {
                 "",
                 "",
                 false
-            )
-            .map(|_| ())
-            .map_err(ErrWrapper),
+            ),
             Err(Error::ConflictingArguments {
                 context: "parse_session_info",
-                args: vec![
-                    "session_hash=09dcee4b212cfd53642ab323fbef07dafafc6f945a80a00147f62910a915c4e6".into(),
-                    "session_name=name".into(),
-                    "session_package_hash=09dcee4b212cfd53642ab323fbef07dafafc6f945a80a00147f62910a915c4e6".into(),
-                    "session_package_name=package_name".into(),
-                    "session_path=./session.wasm".into()
-                ]
-            }
-            .into())
-        );
+                ..
+            })
+        ));
     }
 
     #[test]
     fn should_fail_to_parse_conflicting_payment_parameters() {
-        assert_eq!(
-            parse_payment_info("12345", happy::HASH, happy::NAME, happy::PACKAGE_HASH, happy::PACKAGE_NAME, happy::PATH, &[], "", "", "",)
-                .map(|_| ())
-                .map_err(ErrWrapper),
+        assert!(matches!(
+            parse_payment_info(
+                "12345",
+                happy::HASH,
+                happy::NAME,
+                happy::PACKAGE_HASH,
+                happy::PACKAGE_NAME,
+                happy::PATH,
+                &[],
+                "",
+                "",
+                "",
+            ),
             Err(Error::ConflictingArguments {
                 context: "parse_payment_info",
-                args: vec![
-                    "payment_amount=12345".into(),
-                    "payment_hash=09dcee4b212cfd53642ab323fbef07dafafc6f945a80a00147f62910a915c4e6".into(),
-                    "payment_name=name".into(),
-                    "payment_package_hash=09dcee4b212cfd53642ab323fbef07dafafc6f945a80a00147f62910a915c4e6".into(),
-                    "payment_package_name=package_name".into(),
-                    "payment_path=./session.wasm".into(),
-                ]
-            }
-            .into())
+                ..
+            })
+        ));
+    }
+
+    #[test]
+    fn should_parse_valid_deploy_params() {
+        // create secret key file in tempdir.
+        let keys_dir = tempdir().expect("Failed to create temp dir.");
+        let secret_key_path = keys_dir.path().join("key.pem");
+        let secret_key_path_str = secret_key_path.to_str().unwrap();
+        let mut secret_key_file =
+            fs::File::create(&secret_key_path).expect("Failed to create test secret key file.");
+        write!(secret_key_file, "{}", happy::PRIVATE_KEY)
+            .expect("Failed to write data to test secret key file.");
+
+        // create a valid timestamp and convert it to &str.
+        let timestamp = Timestamp::now().to_string();
+        let timestamp = timestamp.as_str();
+
+        let params = parse_deploy_params(
+            secret_key_path_str,
+            timestamp,
+            "2sec",
+            "10000",
+            &[happy::HASH],
+            "test",
         );
+
+        assert!(params.is_ok());
+    }
+
+    #[test]
+    fn should_fail_to_parse_invalid_deploy_params() {
+        // create secret key file in tempdir.
+        let keys_dir = tempdir().expect("Failed to create temp dir.");
+        let secret_key_path = keys_dir.path().join("key.pem");
+        let secret_key_path_str = secret_key_path.to_str().unwrap();
+        let mut secret_key_file =
+            fs::File::create(&secret_key_path).expect("Failed to create test secret key file.");
+        write!(secret_key_file, "{}", happy::PRIVATE_KEY)
+            .expect("Failed to write data to test secret key file.");
+
+        // create a valid timestamp and convert it to &str.
+        let timestamp = Timestamp::now().to_string();
+        let timestamp = timestamp.as_str();
+
+        let result = parse_deploy_params("bad file path", timestamp, "2sec", "10000", &[], "test");
+
+        // failed to parse secret key file path.
+        assert!(matches!(
+            result.err().expect("Result should be an Err."),
+            Error::CryptoError { .. }
+        ));
+
+        let result = parse_deploy_params(
+            secret_key_path_str,
+            "bad timestamp",
+            "2sec",
+            "10000",
+            &[],
+            "test",
+        );
+
+        // failed to parse timestamp.
+        assert!(matches!(
+            result.err().expect("Result should be an Err."),
+            Error::FailedToParseTimestamp(_, _)
+        ));
+
+        let result = parse_deploy_params(
+            secret_key_path_str,
+            timestamp,
+            "bad ttl",
+            "10000",
+            &[],
+            "test",
+        );
+
+        // failed to parse ttl.
+        assert!(matches!(
+            result.err().expect("Result should be an Err."),
+            Error::FailedToParseTimeDiff(_, _)
+        ));
+
+        let result = parse_deploy_params(
+            secret_key_path_str,
+            timestamp,
+            "2sec",
+            "bad gas price",
+            &[],
+            "test",
+        );
+
+        // failed to parse gas price.
+        assert!(matches!(
+            result.err().expect("Result should be an Err."),
+            Error::FailedToParseInt(_, _)
+        ));
+
+        let result = parse_deploy_params(
+            secret_key_path_str,
+            timestamp,
+            "2sec",
+            "10000",
+            &["bad deploy hash"],
+            "test",
+        );
+
+        // failed to parse deploy hash.
+        assert!(matches!(
+            result.err().expect("Result should be an Err."),
+            Error::CryptoError { .. }
+        ));
     }
 
     mod missing_args {
@@ -974,24 +1054,13 @@ mod tests {
             ($t:ident, $name:ident, $field:tt => $value:expr, missing: $missing:expr, context: $context:expr) => {
                 #[test]
                 fn $name() {
-                    let info: StdResult<ExecutableDeployItem, ErrWrapper> = $t {
+                    let info: StdResult<ExecutableDeployItem, Error> = $t {
                         $field: $value,
                         ..Default::default()
                     }
-                    .try_into()
-                    .map_err(ErrWrapper);
-                    assert_eq!(
-                        info,
-                        Err(Error::InvalidArgument(
-                            $context,
-                            format!(
-                                "Field {} also requires following fields to be provided: {:?}",
-                                stringify!($field),
-                                $missing
-                            )
-                        )
-                        .into())
-                    );
+                    .try_into();
+
+                    assert!(matches!(info, Err(Error::InvalidArgument($context, _msg))));
                 }
             };
         }
@@ -1086,25 +1155,24 @@ mod tests {
         ///
         ///     #[test]
         ///     fn path_conflicts_with_package_hash() {
-        ///         let info: StdResult<ExecutableDeployItem, ErrWrapper> = SessionStrParams {
+        ///         let info: StdResult<ExecutableDeployItem, _> = SessionStrParams {
         ///                 session_path: happy::PATH,
         ///                 session_package_hash: happy::PACKAGE_HASH,
         ///                 ..Default::default()
         ///             }
-        ///             .try_into()
-        ///             .map_err(ErrWrapper);
+        ///             .try_into();
         ///         let mut conflicting = vec![
         ///             format!("{}={}", "session_path", happy::PATH),
         ///             format!("{}={}", "session_package_hash", happy::PACKAGE_HASH),
         ///         ];
         ///         conflicting.sort();
-        ///         assert_eq!(
+        ///         assert!(matches!(
         ///             info,
         ///             Err(Error::ConflictingArguments {
         ///                 context: "parse_session_info",
         ///                 args: conflicting
         ///             }
-        ///             .into())
+        ///             ))
         ///         );
         ///     }
         /// }
@@ -1138,26 +1206,25 @@ mod tests {
                     $(
                         #[test]
                         fn $test_fn_name() {
-                            let info: StdResult<ExecutableDeployItem, ErrWrapper> = $t {
+                            let info: StdResult<ExecutableDeployItem, _> = $t {
                                 $arg: $arg_value,
                                 $con: $con_value,
                                 $($req: $req_value,),*
                                 ..Default::default()
                             }
-                            .try_into()
-                            .map_err(ErrWrapper);
+                            .try_into();
                             let mut conflicting = vec![
                                 format!("{}={}", stringify!($arg), $arg_value),
                                 format!("{}={}", stringify!($con), $con_value),
                             ];
                             conflicting.sort();
-                            assert_eq!(
+                            assert!(matches!(
                                 info,
                                 Err(Error::ConflictingArguments {
                                     context: $context,
-                                    args: conflicting
+                                    ..
                                 }
-                                .into())
+                                ))
                             );
                         }
                     )+