+++ smart_contracts/contract_as/CHANGELOG.md
@@ -13,6 +13,19 @@ All notable changes to this project will be documented in this file.  The format
 
 ## [Unreleased]
 
+### Added
+* Add function to create an account hash from a public key.
+* Add getter for public key algorithm name.
+* Add getter for raw bytes of public key.
+* Export casper_blake2b.
+
+### Fixed
+* Fix instances of `return <T>unreachable()` which is invalid code under the latest AssemblyScript compiler.
+
+
+
+## [1.3.0] - 2021-07-19
+
 No changes.
 
 
@@ -50,7 +63,8 @@ No changes.
 
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0
-[unreleased]: https://github.com/casper-network/casper-node/compare/v1.2.0...dev
+[unreleased]: https://github.com/casper-network/casper-node/compare/v1.3.0...dev
+[1.3.0]: https://github.com/casper-network/casper-node/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/casper-network/casper-node/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/casper-network/casper-node/compare/v1.0.1...v1.1.1
 [1.1.0]: https://github.com/casper-network/casper-node/compare/v1.0.1...v1.1.1